refactor(filid): split review-manage into handlers organ (T1-D)
Split mcp/tools/review-manage/review-manage.ts (564 lines — the only file
exceeding the 500-line FCA hard threshold) into a handlers/ organ plus a
new utils/content-hash.ts shared helper.

Resolves the single hard 500-line violation identified in the ralplan plan.

utils/content-hash.ts (NEW — V1 fix from architect review):
- gitExec            — shared git command execution
- computeContentHash — shared by checkpoint and check-cache handlers

handlers/ (new organ):
- normalize-branch.ts, ensure-dir.ts, checkpoint.ts, elect-committee.ts
- cleanup.ts, content-hash.ts (action dispatcher), check-cache.ts

review-manage.ts is now a 76-line slim dispatcher routing each action to
its handler. handleReviewManage export preserved so src/index.ts surface
is unchanged.

V1 fix applied: computeContentHash lives in utils/, NOT handlers/. It is
a shared utility consumed by multiple handlers, not an action handler.

- src/index.ts diff: empty
- review-manage.ts: 564 -> 76 lines (hard 500-line violation fixed)
- INTENT.md: 38 lines (<=50)
- bundle delta: +0.39% overall
- tests: 846/850 (4 pre-existing context-injector failures)
